requirements.txt
numpy==1.22
matplotlib==3.5
pytest==7.1
pytest-cov==3.0
jupyter==1.0
scipy==1.8.1
astropy==5.1
sunpy==4.0
seaborn==0.11
xarray==2022.6.0
netCDF4==1.6.0
h5netcdf==1.1.0
bottleneck==1.3.7
pillow-jpls==1.2.0
imageio-ffmpeg==2.19.3